refactor: move NMStatusConverter inside nm.status
