+++ kura/org.eclipse.kura.nm/src/main/java/org/eclipse/kura/nm/configuration/NMDbusConnector.java
@@ -22,6 +22,7 @@ import java.util.Optional;
 
 import org.eclipse.kura.net.NetInterface;
 import org.eclipse.kura.net.NetInterfaceAddress;
+import org.eclipse.kura.nm.status.NMStatusConverter;
 import org.freedesktop.NetworkManager;
 import org.freedesktop.dbus.DBusPath;
 import org.freedesktop.dbus.connections.impl.DBusConnection;
